+++ src/Tagged/Html/Embed/Audio.php
@@ -4,27 +4,27 @@
  * @license http://opensource.org/licenses/MIT
  */
 declare(strict_types=1);
-namespace DecodeLabs\Tagged\Embed;
+namespace DecodeLabs\Tagged\Html\Embed;
 
-use DecodeLabs\Tagged\Buffer;
-use DecodeLabs\Tagged\Builder\Html\ContentCollection;
-use DecodeLabs\Tagged\Builder\Html\Tag;
-use DecodeLabs\Tagged\Builder\Html\Element;
-
-use DecodeLabs\Tagged\Embed\Media;
 use DecodeLabs\Tagged\Markup;
+use DecodeLabs\Tagged\Buffer;
 
-class Video implements Media
+use DecodeLabs\Tagged\Html\ContentCollection;
+use DecodeLabs\Tagged\Html\Tag;
+use DecodeLabs\Tagged\Html\Element;
+
+use DecodeLabs\Tagged\Html\Embed\Media;
+use DecodeLabs\Tagged\Html\Embed\MediaTrait;
+
+class Audio implements Media
 {
     use MediaTrait;
 
     const URL_MAP = [
-        'youtube' => 'youtube',
-        'youtu.be' => 'youtube',
-        'vimeo' => 'vimeo'
+        'audioboom' => 'audioboom',
+        'audioboo' => 'audioboom'
     ];
 
-    protected $useApi = false;
 
 
     /**
@@ -32,24 +32,7 @@ class Video implements Media
      */
     public static function defaultUrlFromId(string $id): string
     {
-        return '//www.youtube.com/embed/'.$id;
-    }
-
-    /**
-     * Set to use API (youtube)
-     */
-    public function setUseApi(bool $flag): Media
-    {
-        $this->useApi = $flag;
-        return $this;
-    }
-
-    /**
-     * Should use API?
-     */
-    public function shouldUseApi(): bool
-    {
-        return $this->useApi;
+        return '//embeds.audioboom.com/boos/'.$id.'/embed/v4';
     }
 
 
@@ -60,7 +43,7 @@ class Video implements Media
     public function render(): Markup
     {
         if (($this->url === null || !$this->provider) && $this->source !== null) {
-            return Element::create('div.embed.video', new Buffer($this->source));
+            return Element::create('div.embed.audio', new Buffer($this->source));
         }
 
         return $this->prepareIframeElement($this->url);
@@ -71,7 +54,7 @@ class Video implements Media
      */
     protected function prepareIframeElement(string $url): Element
     {
-        $tag = Element::create('iframe.embed.video', null, [
+        $tag = Element::create('iframe.embed.audio', null, [
             'id' => $this->id,
             'src' => $url,
             'width' => $this->width,
@@ -90,7 +73,7 @@ class Video implements Media
         }
 
         if ($this->provider) {
-            $tag->setAttribute('data-video', $this->provider);
+            $tag->setAttribute('data-audio', $this->provider);
         }
 
         return $tag;