+++ src/Tagged/Html/Embed/MediaTrait.php
@@ -4,7 +4,9 @@
  * @license http://opensource.org/licenses/MIT
  */
 declare(strict_types=1);
-namespace DecodeLabs\Tagged\Embed;
+namespace DecodeLabs\Tagged\Html\Embed;
+
+use DecodeLabs\Tagged\Html\Embed\Media;
 
 use DecodeLabs\Glitch;
 
@@ -130,10 +132,10 @@ trait MediaTrait
      */
     public static function getClassForUrl(string $url): string
     {
-        $class = Video::class;
+        $class = get_called_class();
 
         if ($provider = self::extractProviderName($url)) {
-            $customClass = '\\DecodeLabs\\Tagged\\Embed\\'.ucfirst($provider);
+            $customClass = '\\DecodeLabs\\Tagged\\Html\\Embed\\'.ucfirst($provider);
 
             if (class_exists($customClass)) {
                 $class = $customClass;