+++ src/Tagged/Html/Plugins/Parse.php
@@ -4,13 +4,13 @@
  * @license http://opensource.org/licenses/MIT
  */
 declare(strict_types=1);
-namespace DecodeLabs\Tagged\FactoryPlugins;
+namespace DecodeLabs\Tagged\Html\Plugins;
 
 use DecodeLabs\Veneer\FacadePlugin;
 
 use DecodeLabs\Tagged\Markup;
-use DecodeLabs\Tagged\HtmlFactory;
 use DecodeLabs\Tagged\Buffer;
+use DecodeLabs\Tagged\Html\Factory as HtmlFactory;
 
 use DecodeLabs\Glitch;